feat: 🚀 Create CardExternalLink Component
