+++ app/components/CardExternalLink/CardExternalLink.stories.ts
@@ -1,0 +1,24 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import { CardExternalLink } from "./CardExternalLink";
+
+type T = typeof CardExternalLink;
+
+const meta: Meta = {
+	component: CardExternalLink,
+	parameters: {
+		layout: "centered",
+	},
+	tags: ["autodocs"],
+	title: "CardExternalLink",
+} satisfies Meta<T>;
+
+type Story = StoryObj<T>;
+
+export const Default: Story = {
+	args: {
+		url: "/",
+	},
+};
+
+export default meta;
