+++ app/components/CardExternalLink/index.ts
@@ -1,0 +1,1 @@
+export { CardExternalLink } from "./CardExternalLink";
