+++ app/routes/top/features/DetailCard/DetailCard.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from "react";
 import * as styles from "./styles.css";
-import RightArrow from "~icons/ic/round-play-arrow";
+import { CardExternalLink } from "@/components/CardExternalLink";
 
 interface Props {
 	children: string;
@@ -13,17 +13,7 @@ export function DetailCard({ children, title, url }: Props): ReactNode {
 		<div className={styles.box}>
 			<h3 className={styles.title}>{title}</h3>
 			<p className={styles.content}>{children}</p>
-			<a
-				className={styles.link}
-				href={url}
-				rel="noreferrer"
-				target="_blank"
-			>
-				<span className={styles.span}>
-					本校 HP はこちら
-				</span>
-				<RightArrow className={styles.arrow} />
-			</a>
+			<CardExternalLink url={url} />
 		</div>
 	);
 }
